# src/api.yaml
openapi: 3.0.3
info:
  title: HD1 (Holodeck One) API
  description: |
    Revolutionary API-first 3D coordinate system with universal world boundaries.
    Single source of truth for sessions, worlds, objects, and real-time control.
    
    **CORE PRINCIPLES:**
    - Everything API-driven (zero shell commands)
    - Universal coordinate system [-12, +12] on all axes
    - Objects honor world boundaries absolutely
    - Named object persistence with full lifecycle
    - Real-time WebSocket + REST hybrid
    
    **ARCHITECTURE:**
    - Specification drives code generation
    - Build fails if handlers missing
    - Zero manual routing - auto-generated from spec
    
  version: 1.0.0
  contact:
    name: Holodeck One 3D Engine
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
...
  /channels:
    get:
      operationId: listChannels
      summary: List available collaboration channels
      description: |
        Returns a list of all available collaboration channels with their
        configurations, environments, and capacity information.
      responses:
        '200':
          description: Available channels retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Available channels retrieved successfully"
                  channels:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "channel_one"
                        name:
                          type: string
                          example: "Channel One - Main Collaboration"
                        description:
                          type: string
                          example: "Primary collaborative environment for general holodeck activities"
                        environment:
                          type: string
                          example: "earth-surface"
                        max_clients:
                          type: integer
                          example: 100
      x-handler: "api/channels/list_channels.go"
      x-function: "ListChannels"

    post:
      operationId: createChannel
      summary: Create a new collaboration channel
      description: |
        Creates a new collaboration channel with specified configuration.
        Channel ID will be auto-generated if not provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Channel ID (auto-generated if not provided)
                  example: "channel_custom"
                name:
                  type: string
                  description: Human-readable channel name
                  example: "Custom Collaboration Channel"
                description:
                  type: string
                  description: Channel description and purpose
                  example: "Custom environment for specialized workflows"
                environment:
                  type: string
                  description: Environment ID to apply to channel
                  example: "earth-surface"
                max_clients:
                  type: integer
                  description: Maximum concurrent clients allowed
                  example: 50
                enabled:
                  type: boolean
                  description: Whether channel is active
                  default: true
                priority:
                  type: integer
                  description: Channel priority for display ordering
                  default: 100
              required: [name, environment]
      responses:
        '201':
          description: Channel created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Channel created successfully"
                  channel:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "channel_custom"
                      name:
                        type: string
                        example: "Custom Collaboration Channel"
                      description:
                        type: string
                        example: "Custom environment for specialized workflows"
                      environment:
                        type: string
                        example: "earth-surface"
                      max_clients:
                        type: integer
                        example: 50
                      enabled:
                        type: boolean
                        example: true
                      created_at:
                        type: string
                        format: date-time
        '400':
          description: Invalid channel configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid channel configuration"
                  errors:
                    type: array
                    items:
                      type: string
        '409':
          description: Channel ID already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Channel ID already exists"
      x-handler: "api/channels/create_channel.go"
      x-function: "CreateChannel"
  /channels/{channelId}:
    get:
      operationId: getChannel
      summary: Get channel configuration
      description: |
        Returns detailed configuration for a specific channel including
        current client count and channel statistics.
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          example: "channel_one"
      responses:
        '200':
          description: Channel configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  channel:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "channel_one"
                      name:
                        type: string
                        example: "Channel One - Main Collaboration"
                      description:
                        type: string
                        example: "Primary collaborative environment for general holodeck activities"
                      environment:
                        type: string
                        example: "earth-surface"
                      max_clients:
                        type: integer
                        example: 100
                      enabled:
                        type: boolean
                        example: true
                      current_clients:
                        type: integer
                        example: 5
                      created_at:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Channel not found"
      x-handler: "api/channels/get_channel.go"
      x-function: "GetChannel"

    put:
      operationId: updateChannel
      summary: Update channel configuration
      description: |
        Updates configuration for an existing channel. Only provided fields
        will be updated. Cannot change channel ID.
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          example: "channel_one"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Human-readable channel name
                  example: "Updated Channel Name"
                description:
                  type: string
                  description: Channel description and purpose
                  example: "Updated channel description"
                environment:
                  type: string
                  description: Environment ID to apply to channel
                  example: "molecular-scale"
                max_clients:
                  type: integer
                  description: Maximum concurrent clients allowed
                  example: 75
                enabled:
                  type: boolean
                  description: Whether channel is active
                  example: true
                priority:
                  type: integer
                  description: Channel priority for display ordering
                  example: 50
      responses:
        '200':
          description: Channel updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Channel updated successfully"
                  channel:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "channel_one"
                      name:
                        type: string
                        example: "Updated Channel Name"
                      description:
                        type: string
                        example: "Updated channel description"
                      environment:
                        type: string
                        example: "molecular-scale"
                      max_clients:
                        type: integer
                        example: 75
                      enabled:
                        type: boolean
                        example: true
                      updated_at:
                        type: string
                        format: date-time
        '400':
          description: Invalid channel configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid channel configuration"
                  errors:
                    type: array
                    items:
                      type: string
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Channel not found"
      x-handler: "api/channels/update_channel.go"
      x-function: "UpdateChannel"

    delete:
      operationId: deleteChannel
      summary: Delete a collaboration channel
      description: |
        Deletes an existing channel. Will disconnect all current clients
        and remove channel configuration. Cannot delete default channel.
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          example: "channel_custom"
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Force deletion even if clients are connected
      responses:
        '200':
          description: Channel deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Channel deleted successfully"
                  disconnected_clients:
                    type: integer
                    example: 3
        '400':
          description: Cannot delete channel
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Cannot delete default channel or channel has active clients"
                  current_clients:
                    type: integer
                    example: 5
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Channel not found"
      x-handler: "api/channels/delete_channel.go"
      x-function: "DeleteChannel"

  # ========================================
  # API INFORMATION
  # ========================================
...